% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter_transforms.R
\name{A_from_par}
\alias{A_from_par}
\title{Extract A coefficients}
\usage{
A_from_par(par)
}
\arguments{
\item{par}{Parameter vector with attributes}
}
\value{
A coefficient values
}
\description{
Extract A coefficients
}
