% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter_transforms.R
\name{B_from_par}
\alias{B_from_par}
\title{Extract B coefficients}
\usage{
B_from_par(par)
}
\arguments{
\item{par}{Parameter vector with attributes}
}
\value{
B coefficient values
}
\description{
Extract B coefficients
}
