% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter_transforms.R
\name{calculate_expected_length}
\alias{calculate_expected_length}
\title{Calculate expected parameter vector length for given configuration}
\usage{
calculate_expected_length(K, model_type, diag_probs, equal_variances)
}
\arguments{
\item{K}{Number of regimes}

\item{model_type}{Model type}

\item{diag_probs}{Diagonal transition probabilities flag}

\item{equal_variances}{Equal variances flag}
}
\value{
Expected parameter vector length
}
\description{
Calculate expected parameter vector length for given configuration
}
