% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_functions.R
\name{calculate_fisher_information_diagonal}
\alias{calculate_fisher_information_diagonal}
\title{Calculate Fisher Information for diagonal parameterization}
\usage{
calculate_fisher_information_diagonal(
  mu,
  sigma2,
  X_t_prev,
  p_diag,
  gh_setup,
  K
)
}
\arguments{
\item{mu}{Vector of regime means (length K)}

\item{sigma2}{Vector of regime variances (length K)}

\item{X_t_prev}{Previous filtered probabilities (length K)}

\item{p_diag}{Diagonal transition probabilities (p11, p22, ..., pKK)}

\item{gh_setup}{Gauss-Hermite quadrature setup}

\item{K}{Number of regimes}
}
\value{
Fisher Information scalar
}
\description{
Calculate Fisher Information for diagonal parameterization
}
\details{
Calculates Fisher Information for the diagonal parameterization case.
This follows the methodology from the original HMMGAS C code.
}
