% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transition_helpers.R
\name{compute_initial_predicted_probs}
\alias{compute_initial_predicted_probs}
\title{Compute initial predicted probabilities using stationary distribution}
\usage{
compute_initial_predicted_probs(trans_prob, diag_probs = TRUE)
}
\arguments{
\item{trans_prob}{Vector of transition probabilities}

\item{diag_probs}{If TRUE, trans_prob contains diagonal elements}
}
\value{
Vector of initial predicted probabilities (length K)
}
\description{
Compute initial predicted probabilities using stationary distribution
}
\details{
For HMM initialization, computes X_tlag[1] = P^T \%*\% stationary_dist.
This gives the probability of being in each state at t=1 given we start
from the stationary distribution.

Works for any K >= 2 by using eigenvalue-based stationary distribution
calculation.
}
\examples{
# K=2 example
trans_prob <- c(0.9, 0.8)  # p11, p22
X_tlag_init <- compute_initial_predicted_probs(trans_prob, diag_probs = TRUE)

# K=3 example
trans_prob <- c(0.8, 0.7, 0.9)  # p11, p22, p33
X_tlag_init <- compute_initial_predicted_probs(trans_prob, diag_probs = TRUE)
}
