% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpp_backend.R
\name{cpp_available}
\alias{cpp_available}
\title{Check if C filtering backend is available}
\usage{
cpp_available()
}
\value{
Logical scalar: TRUE if C backend is available, FALSE otherwise.
}
\description{
Check if C filtering backend is available
}
