% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{format_time}
\alias{format_time}
\title{Format execution time in a human-readable way}
\usage{
format_time(seconds)
}
\arguments{
\item{seconds}{Time in seconds}
}
\value{
String with formatted time
}
\description{
Format execution time in a human-readable way
}
\examples{
format_time(65)  # Returns "1m 5s"
}
