% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transition_helpers.R
\name{infer_K_from_probs}
\alias{infer_K_from_probs}
\title{Determine number of regimes from parameter vector length}
\usage{
infer_K_from_probs(probs, diag_probs = TRUE)
}
\arguments{
\item{probs}{Vector of probability parameters}

\item{diag_probs}{Whether parameters use diagonal parameterization}
}
\value{
Number of regimes K
}
\description{
Determine number of regimes from parameter vector length
}
