% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallel_utils.R
\name{is_coordinated_parallel}
\alias{is_coordinated_parallel}
\title{Check if coordinated parallelization will be used}
\usage{
is_coordinated_parallel(allocation)
}
\arguments{
\item{allocation}{Result from get_parallel_allocation()}
}
\value{
TRUE if coordinated, FALSE if estimation-only or sequential
}
\description{
Check if coordinated parallelization will be used
}
\details{
Helper function to check parallelization strategy from allocation result.
Useful for conditional logic in simulation functions.
}
\examples{
allocation <- get_parallel_allocation(6, 50, 3)
if (is_coordinated_parallel(allocation)) {
  # Set up simulation-level parallelization
} else {
  # Use estimation-only parallelization
}
}
