% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallel_utils.R
\name{is_parallel}
\alias{is_parallel}
\title{Check if any parallelization will be used}
\usage{
is_parallel(allocation)
}
\arguments{
\item{allocation}{Result from get_parallel_allocation()}
}
\value{
TRUE if parallel, FALSE if sequential
}
\description{
Check if any parallelization will be used
}
\details{
Helper function to check if any parallel processing will occur.
}
\examples{
\dontrun{
allocation <- get_parallel_allocation(1, 10, 3)
is_parallel(allocation)
}
}
