% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{mad_value}
\alias{mad_value}
\title{Calculate median absolute deviation (robust measure of dispersion)}
\usage{
mad_value(x, na.rm = TRUE, constant = 1.4826)
}
\arguments{
\item{x}{Numeric vector}

\item{na.rm}{Whether to remove NA values (default: TRUE)}

\item{constant}{Scaling factor for normal distribution (default: 1.4826)}
}
\value{
Median absolute deviation
}
\description{
Calculate median absolute deviation (robust measure of dispersion)
}
\examples{
\dontrun{
mad_value(c(1, 2, 3, 100))  # Less affected by outliers than standard deviation
}
}
