% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter_transforms.R
\name{mean_from_par}
\alias{mean_from_par}
\title{Extract mean parameters}
\usage{
mean_from_par(par)
}
\arguments{
\item{par}{Parameter vector with attributes}
}
\value{
Mean parameter values
}
\description{
Extract mean parameters
}
