% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{moving_average}
\alias{moving_average}
\title{Calculate moving average of a time series}
\usage{
moving_average(x, window_size, align = "center")
}
\arguments{
\item{x}{Numeric vector}

\item{window_size}{Number of periods to include in the moving average}

\item{align}{Alignment ("left", "center", "right") (default: "center")}
}
\value{
Vector of moving averages
}
\description{
Calculate moving average of a time series
}
\examples{
moving_average(1:10, 3)  # 3-period centered moving average
}
