% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter_transforms.R
\name{print_parameter_info}
\alias{print_parameter_info}
\title{Print parameter vector structure information}
\usage{
print_parameter_info(par)
}
\arguments{
\item{par}{Parameter vector with attributes}
}
\description{
Print parameter vector structure information
}
