% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transition_helpers.R
\name{print_transition_matrix}
\alias{print_transition_matrix}
\title{Print transition matrix in a readable format}
\usage{
print_transition_matrix(t_mat, digits = 4)
}
\arguments{
\item{t_mat}{Transition matrix}

\item{digits}{Number of decimal places to display (default: 4)}
}
\description{
Print transition matrix in a readable format
}
