% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{progress_bar}
\alias{progress_bar}
\title{Create a progress bar for lengthy operations}
\usage{
progress_bar(total, title = "Progress", width = 50)
}
\arguments{
\item{total}{Total number of iterations}

\item{title}{Title for the progress bar (default: "Progress")}

\item{width}{Width of the progress bar (default: 50)}
}
\value{
Function that updates the progress bar
}
\description{
Create a progress bar for lengthy operations
}
\examples{
n <- 100
pb <- progress_bar(n, "Processing")
for (i in 1:n) {
  # Do some work
  Sys.sleep(0.01)
  pb(i)  # Update progress bar
}
}
