% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{replace_invalid_values}
\alias{replace_invalid_values}
\title{Replace invalid values (NA, NaN, Inf, -Inf) with a default value}
\usage{
replace_invalid_values(x, replacement = 0)
}
\arguments{
\item{x}{Vector to process}

\item{replacement}{Value to use for replacements (default: 0)}
}
\value{
Vector with invalid values replaced
}
\description{
Replace invalid values (NA, NaN, Inf, -Inf) with a default value
}
\examples{
\dontrun{
replace_invalid_values(c(1, 2, NA, Inf, -Inf, 6))  # Returns c(1, 2, 0, 0, 0, 6)
}
}
