% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter_transforms.R
\name{sigma2_from_par}
\alias{sigma2_from_par}
\title{Extract variance parameters}
\usage{
sigma2_from_par(par)
}
\arguments{
\item{par}{Parameter vector with attributes}
}
\value{
Variance parameter values
}
\description{
Extract variance parameters
}
