% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transition_helpers.R
\name{transition_matrix_offdiagonal}
\alias{transition_matrix_offdiagonal}
\title{Create transition matrix from off-diagonal parameterization}
\usage{
transition_matrix_offdiagonal(off_diag_probs, check_validity = TRUE)
}
\arguments{
\item{off_diag_probs}{Vector of off-diagonal transition probabilities}

\item{check_validity}{Logical; whether to validate the resulting matrix}
}
\value{
K x K transition matrix
}
\description{
Create transition matrix from off-diagonal parameterization
}
\details{
Creates a transition matrix from K*(K-1) off-diagonal elements specified
in row-major order. Diagonal elements are calculated as p_ii = 1 - sum(p_ij for j≠i).

This is the default parameterization for the new implementation.
}
