% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter_transforms.R
\name{transp_from_par}
\alias{transp_from_par}
\title{Extract transition probabilities}
\usage{
transp_from_par(par)
}
\arguments{
\item{par}{Parameter vector with attributes}
}
\value{
Transition probability values
}
\description{
Extract transition probabilities
}
