% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter_transforms.R
\name{untransform_parameters}
\alias{untransform_parameters}
\title{Transform parameters from unconstrained space back to natural space}
\usage{
untransform_parameters(par_t)
}
\arguments{
\item{par_t}{Parameter vector with attributes in transformed space}
}
\value{
Parameter vector with attributes in natural space
}
\description{
Transform parameters from unconstrained space back to natural space
}
