% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_functions.R
\name{validate_gauss_hermite_setup}
\alias{validate_gauss_hermite_setup}
\title{Validate Gauss-Hermite setup object}
\usage{
validate_gauss_hermite_setup(gh_setup)
}
\arguments{
\item{gh_setup}{A gauss_hermite_setup object to validate}
}
\value{
TRUE if valid, otherwise throws an error with explanation
}
\description{
Validate Gauss-Hermite setup object
}
\details{
Checks that the Gauss-Hermite setup object has all required components
and that they are valid for use in score calculations.
}
\examples{
\dontrun{
y <- rnorm(1000)
gh_setup <- setup_gauss_hermite_quadrature(y)
validate_gauss_hermite_setup(gh_setup)  # Returns TRUE
}
}
