% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter_transforms.R
\name{validate_parameter_attributes}
\alias{validate_parameter_attributes}
\title{Validate parameter vector has required attributes}
\usage{
validate_parameter_attributes(par)
}
\arguments{
\item{par}{Parameter vector to validate}
}
\value{
TRUE if valid, throws error if not
}
\description{
Validate parameter vector has required attributes
}
